+++ src/features/auth/components/index.ts
@@ -1,0 +1,1 @@
+export { default as SignUpField } from './sign-up-field/SignUpField'
