+++ src/features/auth/components/sign-up-field/SignUpField.type.ts
@@ -1,0 +1,6 @@
+export interface SignUpFormData {
+  email: string
+  nickname: string
+  password: string
+  confirmPassword: string
+}
